pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "chessformer"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.21.0",
    "python-chess>=1.999",
    "pygame>=2.6.1",
]

[tool.setuptools]
py-modules = ["chessformer"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
